refactor: Split files in strategy crate into modules
