+++ crates/strategy/zksync/src/cheatcode/runner/mod.rs
@@ -1,50 +1,31 @@
-use std::{any::TypeId, fs, path::PathBuf, sync::Arc};
+use std::sync::Arc;
 
-use alloy_json_abi::ContractObject;
-use alloy_primitives::{keccak256, map::HashMap, Address, Bytes, TxKind, B256, U256};
+use alloy_primitives::{map::HashMap, Address, Bytes, TxKind, B256, U256};
 use alloy_rpc_types::{
     request::{TransactionInput, TransactionRequest},
     serde_helpers::WithOtherFields,
 };
-use alloy_sol_types::SolValue;
 use foundry_cheatcodes::{
-    journaled_account, make_acc_non_empty,
+    journaled_account,
     strategy::{
-        CheatcodeInspectorStrategy, CheatcodeInspectorStrategyContext,
-        CheatcodeInspectorStrategyExt, CheatcodeInspectorStrategyRunner,
-        EvmCheatcodeInspectorStrategyRunner,
+        CheatcodeInspectorStrategyContext, CheatcodeInspectorStrategyExt,
+        CheatcodeInspectorStrategyRunner, EvmCheatcodeInspectorStrategyRunner,
     },
     Broadcast, BroadcastableTransaction, BroadcastableTransactions, Cheatcodes, CheatcodesExecutor,
-    CheatsConfig, CheatsCtxt, CommonCreateInput, DealRecord, DynCheatcode, Ecx, Error, InnerEcx,
-    Result,
-    Vm::{
-        self, createFork_0Call, createFork_1Call, createFork_2Call, createSelectFork_0Call,
-        createSelectFork_1Call, createSelectFork_2Call, dealCall, etchCall, getCodeCall,
-        getNonce_0Call, mockCallRevert_0Call, mockCall_0Call, resetNonceCall, rollCall,
-        selectForkCall, setNonceCall, setNonceUnsafeCall, warpCall, zkRegisterContractCall,
-        zkUseFactoryDepCall, zkUsePaymasterCall, zkVmCall, zkVmSkipCall,
-    },
+    CheatsConfig, CheatsCtxt, CommonCreateInput, DynCheatcode, Ecx, InnerEcx, Result, Vm,
 };
 use foundry_common::TransactionMaybeSigned;
-use foundry_config::fs_permissions::FsAccessKind;
 use foundry_evm::{
     backend::{DatabaseError, LocalForkId},
     constants::{DEFAULT_CREATE2_DEPLOYER, DEFAULT_CREATE2_DEPLOYER_CODE},
 };
-use foundry_evm_core::{
-    backend::DatabaseExt,
-    constants::{CHEATCODE_ADDRESS, CHEATCODE_CONTRACT_HASH},
-};
-use foundry_zksync_compilers::dual_compiled_contracts::{
-    ContractType, DualCompiledContract, DualCompiledContracts,
-};
+use foundry_evm_core::backend::DatabaseExt;
 use foundry_zksync_core::{
     convert::{ConvertAddress, ConvertH160, ConvertH256, ConvertRU256, ConvertU256},
-    get_account_code_key, get_balance_key, get_nonce_key,
-    vm::ZkEnv,
-    PaymasterParams, ZkPaymasterData, ZkTransactionMetadata, ACCOUNT_CODE_STORAGE_ADDRESS,
-    CONTRACT_DEPLOYER_ADDRESS, DEFAULT_CREATE2_DEPLOYER_ZKSYNC, H256, KNOWN_CODES_STORAGE_ADDRESS,
-    L2_BASE_TOKEN_ADDRESS, NONCE_HOLDER_ADDRESS, ZKSYNC_TRANSACTION_OTHER_FIELDS_KEY,
+    get_account_code_key, get_balance_key, get_nonce_key, PaymasterParams, ZkTransactionMetadata,
+    ACCOUNT_CODE_STORAGE_ADDRESS, CONTRACT_DEPLOYER_ADDRESS, DEFAULT_CREATE2_DEPLOYER_ZKSYNC,
+    KNOWN_CODES_STORAGE_ADDRESS, L2_BASE_TOKEN_ADDRESS, NONCE_HOLDER_ADDRESS,
+    ZKSYNC_TRANSACTION_OTHER_FIELDS_KEY,
 };
 use itertools::Itertools;
 use revm::{
@@ -57,191 +38,21 @@ use revm::{
         SignedAuthorization, KECCAK_EMPTY,
     },
 };
-use semver::Version;
-use tracing::{debug, error, info, warn};
+use tracing::{debug, error, info};
 use zksync_types::{
     block::{pack_block_info, unpack_block_info},
     utils::{decompose_full_nonce, nonces_to_full_nonce},
     CURRENT_VIRTUAL_BLOCK_INFO_POSITION, SYSTEM_CONTEXT_ADDRESS,
 };
 
-macro_rules! fmt_err {
-    ($msg:literal $(,)?) => {
-        Error::fmt(::std::format_args!($msg))
-    };
-    ($err:expr $(,)?) => {
-        <Error as ::std::convert::From<_>>::from($err)
-    };
-    ($fmt:expr, $($arg:tt)*) => {
-        Error::fmt(::std::format_args!($fmt, $($arg)*))
-    };
-}
+use crate::cheatcode::context::ZksyncCheatcodeInspectorStrategyContext;
 
-macro_rules! bail {
-    ($msg:literal $(,)?) => {
-        return ::std::result::Result::Err(fmt_err!($msg))
-    };
-    ($err:expr $(,)?) => {
-        return ::std::result::Result::Err(fmt_err!($err))
-    };
-    ($fmt:expr, $($arg:tt)*) => {
-        return ::std::result::Result::Err(fmt_err!($fmt, $($arg)*))
-    };
-}
+mod cheatcode_handlers;
+mod utils;
 
 /// ZKsync implementation for [CheatcodeInspectorStrategyRunner].
 #[derive(Debug, Default, Clone)]
 pub struct ZksyncCheatcodeInspectorStrategyRunner;
-
-/// Context for [ZksyncCheatcodeInspectorStrategyRunner].
-#[derive(Debug, Default, Clone)]
-pub struct ZksyncCheatcodeInspectorStrategyContext {
-    pub using_zk_vm: bool,
-
-    /// When in zkEVM context, execute the next CALL or CREATE in the EVM instead.
-    pub skip_zk_vm: bool,
-
-    /// Any contracts that were deployed in `skip_zk_vm` step.
-    /// This makes it easier to dispatch calls to any of these addresses in zkEVM context, directly
-    /// to EVM. Alternatively, we'd need to add `vm.zkVmSkip()` to these calls manually.
-    pub skip_zk_vm_addresses: HashSet<Address>,
-
-    /// Records the next create address for `skip_zk_vm_addresses`.
-    pub record_next_create_address: bool,
-
-    /// Paymaster params
-    pub paymaster_params: Option<ZkPaymasterData>,
-
-    /// Dual compiled contracts
-    pub dual_compiled_contracts: DualCompiledContracts,
-
-    /// The migration status of the database to zkEVM storage, `None` if we start in EVM context.
-    pub zk_startup_migration: ZkStartupMigration,
-
-    /// Factory deps stored through `zkUseFactoryDep`. These factory deps are used in the next
-    /// CREATE or CALL, and cleared after.
-    pub zk_use_factory_deps: Vec<String>,
-
-    /// The list of factory_deps seen so far during a test or script execution.
-    /// Ideally these would be persisted in the storage, but since modifying [revm::JournaledState]
-    /// would be a significant refactor, we maintain the factory_dep part in the [Cheatcodes].
-    /// This can be done as each test runs with its own [Cheatcodes] instance, thereby
-    /// providing the necessary level of isolation.
-    pub persisted_factory_deps: HashMap<H256, Vec<u8>>,
-
-    /// Nonce update persistence behavior in zkEVM for the tx caller.
-    pub zk_persist_nonce_update: ZkPersistNonceUpdate,
-
-    /// Stores the factory deps that were detected as part of CREATE2 deployer call.
-    /// Must be cleared every call.
-    pub set_deployer_call_input_factory_deps: Vec<Vec<u8>>,
-
-    /// Era Vm environment
-    pub zk_env: ZkEnv,
-}
-
-impl ZksyncCheatcodeInspectorStrategyContext {
-    pub fn new(dual_compiled_contracts: DualCompiledContracts, zk_env: ZkEnv) -> Self {
-        // We add the empty bytecode manually so it is correctly translated in zk mode.
-        // This is used in many places in foundry, e.g. in cheatcode contract's account code.
-        let empty_bytes = Bytes::from_static(&[0]);
-        let zk_bytecode_hash = foundry_zksync_core::hash_bytecode(&foundry_zksync_core::EMPTY_CODE);
-        let zk_deployed_bytecode = foundry_zksync_core::EMPTY_CODE.to_vec();
-
-        let mut dual_compiled_contracts = dual_compiled_contracts;
-        dual_compiled_contracts.push(DualCompiledContract {
-            name: String::from("EmptyEVMBytecode"),
-            zk_bytecode_hash,
-            zk_deployed_bytecode: zk_deployed_bytecode.clone(),
-            zk_factory_deps: Default::default(),
-            evm_bytecode_hash: B256::from_slice(&keccak256(&empty_bytes)[..]),
-            evm_deployed_bytecode: Bytecode::new_raw(empty_bytes.clone()).bytecode().to_vec(),
-            evm_bytecode: Bytecode::new_raw(empty_bytes).bytecode().to_vec(),
-        });
-
-        let cheatcodes_bytecode = {
-            let mut bytecode = CHEATCODE_ADDRESS.abi_encode_packed();
-            bytecode.append(&mut [0; 12].to_vec());
-            Bytes::from(bytecode)
-        };
-        dual_compiled_contracts.push(DualCompiledContract {
-            name: String::from("CheatcodeBytecode"),
-            // we put a different bytecode hash here so when importing back to EVM
-            // we avoid collision with EmptyEVMBytecode for the cheatcodes
-            zk_bytecode_hash: foundry_zksync_core::hash_bytecode(CHEATCODE_CONTRACT_HASH.as_ref()),
-            zk_deployed_bytecode: cheatcodes_bytecode.to_vec(),
-            zk_factory_deps: Default::default(),
-            evm_bytecode_hash: CHEATCODE_CONTRACT_HASH,
-            evm_deployed_bytecode: cheatcodes_bytecode.to_vec(),
-            evm_bytecode: cheatcodes_bytecode.to_vec(),
-        });
-
-        let mut persisted_factory_deps = HashMap::new();
-        persisted_factory_deps.insert(zk_bytecode_hash, zk_deployed_bytecode);
-
-        Self {
-            using_zk_vm: false, // We need to migrate once on initialize_interp
-            skip_zk_vm: false,
-            skip_zk_vm_addresses: Default::default(),
-            record_next_create_address: Default::default(),
-            paymaster_params: Default::default(),
-            dual_compiled_contracts,
-            zk_startup_migration: ZkStartupMigration::Defer,
-            zk_use_factory_deps: Default::default(),
-            persisted_factory_deps: Default::default(),
-            zk_persist_nonce_update: Default::default(),
-            set_deployer_call_input_factory_deps: Default::default(),
-            zk_env,
-        }
-    }
-}
-
-impl CheatcodeInspectorStrategyContext for ZksyncCheatcodeInspectorStrategyContext {
-    fn new_cloned(&self) -> Box<dyn CheatcodeInspectorStrategyContext> {
-        Box::new(self.clone())
-    }
-
-    fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
-        self
-    }
-
-    fn as_any_ref(&self) -> &dyn std::any::Any {
-        self
-    }
-}
-
-/// Allows overriding nonce update behavior for the tx caller in the zkEVM.
-///
-/// Since each CREATE or CALL is executed as a separate transaction within zkEVM, we currently skip
-/// persisting nonce updates as it erroneously increments the tx nonce. However, under certain
-/// situations, e.g. deploying contracts, transacts, etc. the nonce updates must be persisted.
-#[derive(Default, Debug, Clone)]
-pub enum ZkPersistNonceUpdate {
-    /// Never update the nonce. This is currently the default behavior.
-    #[default]
-    Never,
-    /// Override the default behavior, and persist nonce update for tx caller for the next
-    /// zkEVM execution _only_.
-    PersistNext,
-}
-
-impl ZkPersistNonceUpdate {
-    /// Persist nonce update for the tx caller for next execution.
-    pub fn persist_next(&mut self) {
-        *self = Self::PersistNext;
-    }
-
-    /// Retrieve if a nonce update must be persisted, or not. Resets the state to default.
-    pub fn check(&mut self) -> bool {
-        let persist_nonce_update = match self {
-            Self::Never => false,
-            Self::PersistNext => true,
-        };
-        *self = Default::default();
-
-        persist_nonce_update
-    }
-}
 
 impl CheatcodeInspectorStrategyRunner for ZksyncCheatcodeInspectorStrategyRunner {
     fn base_contract_deployed(&self, ctx: &mut dyn CheatcodeInspectorStrategyContext) {
@@ -259,289 +70,7 @@ impl CheatcodeInspectorStrategyRunner for ZksyncCheatcodeInspectorStrategyRunner
         ccx: &mut CheatsCtxt<'_, '_, '_, '_>,
         executor: &mut dyn CheatcodesExecutor,
     ) -> Result {
-        fn is<T: std::any::Any>(t: TypeId) -> bool {
-            TypeId::of::<T>() == t
-        }
-
-        let using_zk_vm = get_context(ccx.state.strategy.context.as_mut()).using_zk_vm;
-
-        // Try to downcast the cheatcode to a type that requires special handling.
-        // Note that some cheatcodes are only handled in zkEVM context.
-        // If no handler fires, we use the default execution logic.
-        match cheatcode.as_any().type_id() {
-            t if using_zk_vm && is::<etchCall>(t) => {
-                let etchCall { target, newRuntimeBytecode } =
-                    cheatcode.as_any().downcast_ref().unwrap();
-                foundry_zksync_core::cheatcodes::etch(*target, newRuntimeBytecode, ccx.ecx);
-                Ok(Default::default())
-            }
-            t if using_zk_vm && is::<rollCall>(t) => {
-                let &rollCall { newHeight } = cheatcode.as_any().downcast_ref().unwrap();
-                ccx.ecx.env.block.number = newHeight;
-                foundry_zksync_core::cheatcodes::roll(newHeight, ccx.ecx);
-                Ok(Default::default())
-            }
-            t if using_zk_vm && is::<warpCall>(t) => {
-                let &warpCall { newTimestamp } = cheatcode.as_any().downcast_ref().unwrap();
-                ccx.ecx.env.block.number = newTimestamp;
-                foundry_zksync_core::cheatcodes::warp(newTimestamp, ccx.ecx);
-                Ok(Default::default())
-            }
-            t if using_zk_vm && is::<dealCall>(t) => {
-                let &dealCall { account, newBalance } = cheatcode.as_any().downcast_ref().unwrap();
-
-                let old_balance =
-                    foundry_zksync_core::cheatcodes::deal(account, newBalance, ccx.ecx);
-                let record = DealRecord { address: account, old_balance, new_balance: newBalance };
-                ccx.state.eth_deals.push(record);
-                Ok(Default::default())
-            }
-            t if using_zk_vm && is::<resetNonceCall>(t) => {
-                let &resetNonceCall { account } = cheatcode.as_any().downcast_ref().unwrap();
-                foundry_zksync_core::cheatcodes::set_nonce(account, U256::ZERO, ccx.ecx);
-                Ok(Default::default())
-            }
-            t if using_zk_vm && is::<setNonceCall>(t) => {
-                let &setNonceCall { account, newNonce } =
-                    cheatcode.as_any().downcast_ref().unwrap();
-
-                // nonce must increment only
-                let current = foundry_zksync_core::cheatcodes::get_nonce(account, ccx.ecx);
-                if U256::from(newNonce) < current {
-                    return Err(fmt_err!(
-                        "new nonce ({newNonce}) must be strictly equal to or higher than the \
-                    account's current nonce ({current})"
-                    ));
-                }
-
-                foundry_zksync_core::cheatcodes::set_nonce(account, U256::from(newNonce), ccx.ecx);
-                Ok(Default::default())
-            }
-            t if using_zk_vm && is::<setNonceUnsafeCall>(t) => {
-                let &setNonceUnsafeCall { account, newNonce } =
-                    cheatcode.as_any().downcast_ref().unwrap();
-                foundry_zksync_core::cheatcodes::set_nonce(account, U256::from(newNonce), ccx.ecx);
-                Ok(Default::default())
-            }
-            t if using_zk_vm && is::<getNonce_0Call>(t) => {
-                let &getNonce_0Call { account } = cheatcode.as_any().downcast_ref().unwrap();
-
-                let nonce = foundry_zksync_core::cheatcodes::get_nonce(account, ccx.ecx);
-                Ok(nonce.abi_encode())
-            }
-            t if using_zk_vm && is::<mockCall_0Call>(t) => {
-                let mockCall_0Call { callee, data, returnData } =
-                    cheatcode.as_any().downcast_ref().unwrap();
-
-                let _ = foundry_cheatcodes::make_acc_non_empty(callee, ccx.ecx)?;
-                foundry_zksync_core::cheatcodes::set_mocked_account(*callee, ccx.ecx, ccx.caller);
-                foundry_cheatcodes::mock_call(
-                    ccx.state,
-                    callee,
-                    data,
-                    None,
-                    returnData,
-                    InstructionResult::Return,
-                );
-                Ok(Default::default())
-            }
-            t if using_zk_vm && is::<mockCallRevert_0Call>(t) => {
-                let mockCallRevert_0Call { callee, data, revertData } =
-                    cheatcode.as_any().downcast_ref().unwrap();
-
-                let _ = make_acc_non_empty(callee, ccx.ecx)?;
-                foundry_zksync_core::cheatcodes::set_mocked_account(*callee, ccx.ecx, ccx.caller);
-                // not calling
-                foundry_cheatcodes::mock_call(
-                    ccx.state,
-                    callee,
-                    data,
-                    None,
-                    revertData,
-                    InstructionResult::Revert,
-                );
-                Ok(Default::default())
-            }
-            t if is::<getCodeCall>(t) => {
-                // We don't need to check for `using_zk_vm` since we pass it to `get_artifact_code`.
-                let getCodeCall { artifactPath } = cheatcode.as_any().downcast_ref().unwrap();
-
-                let ctx = get_context(ccx.state.strategy.context.as_mut());
-                Ok(get_artifact_code(
-                    &ctx.dual_compiled_contracts,
-                    ctx.using_zk_vm,
-                    &ccx.state.config,
-                    artifactPath,
-                    false,
-                )?
-                .abi_encode())
-            }
-            t if is::<zkVmCall>(t) => {
-                let zkVmCall { enable } = cheatcode.as_any().downcast_ref().unwrap();
-                let ctx = get_context(ccx.state.strategy.context.as_mut());
-                if *enable {
-                    self.select_zk_vm(ctx, ccx.ecx, None)
-                } else {
-                    self.select_evm(ctx, ccx.ecx);
-                }
-                Ok(Default::default())
-            }
-            t if is::<zkVmSkipCall>(t) => {
-                let zkVmSkipCall { .. } = cheatcode.as_any().downcast_ref().unwrap();
-                let ctx = get_context(ccx.state.strategy.context.as_mut());
-                ctx.skip_zk_vm = true;
-                Ok(Default::default())
-            }
-            t if is::<zkUsePaymasterCall>(t) => {
-                let zkUsePaymasterCall { paymaster_address, paymaster_input } =
-                    cheatcode.as_any().downcast_ref().unwrap();
-                let ctx = get_context(ccx.state.strategy.context.as_mut());
-                ctx.paymaster_params = Some(ZkPaymasterData {
-                    address: *paymaster_address,
-                    input: paymaster_input.clone(),
-                });
-                Ok(Default::default())
-            }
-            t if is::<zkUseFactoryDepCall>(t) => {
-                let zkUseFactoryDepCall { name } = cheatcode.as_any().downcast_ref().unwrap();
-                info!("Adding factory dependency: {:?}", name);
-                let ctx = get_context(ccx.state.strategy.context.as_mut());
-                ctx.zk_use_factory_deps.push(name.clone());
-                Ok(Default::default())
-            }
-            t if is::<zkRegisterContractCall>(t) => {
-                let zkRegisterContractCall {
-                    name,
-                    evmBytecodeHash,
-                    evmDeployedBytecode,
-                    evmBytecode,
-                    zkBytecodeHash,
-                    zkDeployedBytecode,
-                } = cheatcode.as_any().downcast_ref().unwrap();
-                let ctx = get_context(ccx.state.strategy.context.as_mut());
-
-                let zk_factory_deps = vec![]; //TODO: add argument to cheatcode
-                let new_contract = DualCompiledContract {
-                    name: name.clone(),
-                    zk_bytecode_hash: H256(zkBytecodeHash.0),
-                    zk_deployed_bytecode: zkDeployedBytecode.to_vec(),
-                    zk_factory_deps,
-                    evm_bytecode_hash: *evmBytecodeHash,
-                    evm_deployed_bytecode: evmDeployedBytecode.to_vec(),
-                    evm_bytecode: evmBytecode.to_vec(),
-                };
-
-                if let Some(existing) = ctx.dual_compiled_contracts.iter().find(|contract| {
-                    contract.evm_bytecode_hash == new_contract.evm_bytecode_hash &&
-                        contract.zk_bytecode_hash == new_contract.zk_bytecode_hash
-                }) {
-                    warn!(
-                        name = existing.name,
-                        "contract already exists with the given bytecode hashes"
-                    );
-                    return Ok(Default::default())
-                }
-
-                ctx.dual_compiled_contracts.push(new_contract);
-
-                Ok(Default::default())
-            }
-            t if is::<selectForkCall>(t) => {
-                let selectForkCall { forkId } = cheatcode.as_any().downcast_ref().unwrap();
-                let ctx = get_context(ccx.state.strategy.context.as_mut());
-
-                // Re-implementation of `persist_caller` from `fork.rs`.
-                ccx.ecx.db.add_persistent_account(ccx.caller);
-
-                // Prepare storage.
-                self.select_fork_vm(ctx, ccx.ecx, *forkId);
-
-                // Apply cheatcode as usual.
-                cheatcode.dyn_apply(ccx, executor)
-            }
-            t if is::<createSelectFork_0Call>(t) => {
-                let createSelectFork_0Call { urlOrAlias } =
-                    cheatcode.as_any().downcast_ref().unwrap();
-
-                // Re-implementation of `persist_caller` from `fork.rs`.
-                ccx.ecx.db.add_persistent_account(ccx.caller);
-
-                // Create fork.
-                let create_fork_cheatcode = createFork_0Call { urlOrAlias: urlOrAlias.clone() };
-
-                let encoded_fork_id = create_fork_cheatcode.dyn_apply(ccx, executor)?;
-                let fork_id = LocalForkId::abi_decode(&encoded_fork_id, true)?;
-
-                // Prepare storage.
-                {
-                    let ctx = get_context(ccx.state.strategy.context.as_mut());
-                    self.select_fork_vm(ctx, ccx.ecx, fork_id);
-                }
-
-                // Select fork
-                let select_fork_cheatcode = selectForkCall { forkId: fork_id };
-                select_fork_cheatcode.dyn_apply(ccx, executor)?;
-
-                // We need to return the fork ID.
-                Ok(encoded_fork_id)
-            }
-            t if is::<createSelectFork_1Call>(t) => {
-                let createSelectFork_1Call { urlOrAlias, blockNumber } =
-                    cheatcode.as_any().downcast_ref().unwrap();
-
-                // Re-implementation of `persist_caller` from `fork.rs`.
-                ccx.ecx.db.add_persistent_account(ccx.caller);
-
-                // Create fork.
-                let create_fork_cheatcode =
-                    createFork_1Call { urlOrAlias: urlOrAlias.clone(), blockNumber: *blockNumber };
-                let encoded_fork_id = create_fork_cheatcode.dyn_apply(ccx, executor)?;
-                let fork_id = LocalForkId::abi_decode(&encoded_fork_id, true)?;
-
-                // Prepare storage.
-                {
-                    let ctx = get_context(ccx.state.strategy.context.as_mut());
-                    self.select_fork_vm(ctx, ccx.ecx, fork_id);
-                }
-
-                // Select fork
-                let select_fork_cheatcode = selectForkCall { forkId: fork_id };
-                select_fork_cheatcode.dyn_apply(ccx, executor)?;
-
-                // We need to return the fork ID.
-                Ok(encoded_fork_id)
-            }
-            t if is::<createSelectFork_2Call>(t) => {
-                let createSelectFork_2Call { urlOrAlias, txHash } =
-                    cheatcode.as_any().downcast_ref().unwrap();
-
-                // Re-implementation of `persist_caller` from `fork.rs`.
-                ccx.ecx.db.add_persistent_account(ccx.caller);
-
-                // Create fork.
-                let create_fork_cheatcode =
-                    createFork_2Call { urlOrAlias: urlOrAlias.clone(), txHash: *txHash };
-                let encoded_fork_id = create_fork_cheatcode.dyn_apply(ccx, executor)?;
-                let fork_id = LocalForkId::abi_decode(&encoded_fork_id, true)?;
-
-                // Prepare storage.
-                {
-                    let ctx = get_context(ccx.state.strategy.context.as_mut());
-                    self.select_fork_vm(ctx, ccx.ecx, fork_id);
-                }
-
-                // Select fork
-                let select_fork_cheatcode = selectForkCall { forkId: fork_id };
-                select_fork_cheatcode.dyn_apply(ccx, executor)?;
-
-                // We need to return the fork ID.
-                Ok(encoded_fork_id)
-            }
-            _ => {
-                // Not custom, just invoke the default behavior
-                cheatcode.dyn_apply(ccx, executor)
-            }
-        }
+        self.apply_cheatcode_impl(cheatcode, ccx, executor)
     }
 
     fn record_broadcastable_create_transactions(
@@ -603,7 +132,7 @@ impl CheatcodeInspectorStrategyRunner for ZksyncCheatcodeInspectorStrategyRunner
             .zk_use_factory_deps
             .iter()
             .map(|contract| {
-                get_artifact_code(
+                utils::get_artifact_code(
                     &ctx.dual_compiled_contracts,
                     ctx.using_zk_vm,
                     &config,
@@ -705,7 +234,7 @@ impl CheatcodeInspectorStrategyRunner for ZksyncCheatcodeInspectorStrategyRunner
             .zk_use_factory_deps
             .iter()
             .flat_map(|contract| {
-                let artifact_code = get_artifact_code(
+                let artifact_code = utils::get_artifact_code(
                     &ctx.dual_compiled_contracts,
                     ctx.using_zk_vm,
                     &config,
@@ -953,7 +482,7 @@ impl CheatcodeInspectorStrategyExt for ZksyncCheatcodeInspectorStrategyRunner {
             .zk_use_factory_deps
             .iter()
             .flat_map(|contract| {
-                let artifact_code = get_artifact_code(
+                let artifact_code = utils::get_artifact_code(
                     &ctx.dual_compiled_contracts,
                     ctx.using_zk_vm,
                     &state.config,
@@ -1481,189 +1010,6 @@ impl ZksyncCheatcodeInspectorStrategyRunner {
             account.info.code.clone_from(&info.code);
         }
     }
-}
-
-/// Setting for migrating the database to zkEVM storage when starting in ZKsync mode.
-/// The migration is performed on the DB via the inspector so must only be performed once.
-#[derive(Debug, Default, Clone)]
-pub enum ZkStartupMigration {
-    /// Defer database migration to a later execution point.
-    ///
-    /// This is required as we need to wait for some baseline deployments
-    /// to occur before the test/script execution is performed.
-    #[default]
-    Defer,
-    /// Allow database migration.
-    Allow,
-    /// Database migration has already been performed.
-    Done,
-}
-
-impl ZkStartupMigration {
-    /// Check if startup migration is allowed. Migration is disallowed if it's to be deferred or has
-    /// already been performed.
-    pub fn is_allowed(&self) -> bool {
-        matches!(self, Self::Allow)
-    }
-
-    /// Allow migrating the the DB to zkEVM storage.
-    pub fn allow(&mut self) {
-        *self = Self::Allow
-    }
-
-    /// Mark the migration as completed. It must not be performed again.
-    pub fn done(&mut self) {
-        *self = Self::Done
-    }
-}
-
-/// Create ZKsync strategy for [CheatcodeInspectorStrategy].
-pub trait ZksyncCheatcodeInspectorStrategyBuilder {
-    /// Create new ZKsync strategy.
-    fn new_zksync(dual_compiled_contracts: DualCompiledContracts, zk_env: ZkEnv) -> Self;
-}
-
-impl ZksyncCheatcodeInspectorStrategyBuilder for CheatcodeInspectorStrategy {
-    fn new_zksync(dual_compiled_contracts: DualCompiledContracts, zk_env: ZkEnv) -> Self {
-        Self {
-            runner: &ZksyncCheatcodeInspectorStrategyRunner,
-            context: Box::new(ZksyncCheatcodeInspectorStrategyContext::new(
-                dual_compiled_contracts,
-                zk_env,
-            )),
-        }
-    }
-}
-
-fn get_artifact_code(
-    dual_compiled_contracts: &DualCompiledContracts,
-    using_zk_vm: bool,
-    config: &Arc<CheatsConfig>,
-    path: &str,
-    deployed: bool,
-) -> Result<Bytes> {
-    let path = if path.ends_with(".json") {
-        PathBuf::from(path)
-    } else {
-        let mut parts = path.split(':');
-
-        let mut file = None;
-        let mut contract_name = None;
-        let mut version = None;
-
-        let path_or_name = parts.next().unwrap();
-        if path_or_name.contains('.') {
-            file = Some(PathBuf::from(path_or_name));
-            if let Some(name_or_version) = parts.next() {
-                if name_or_version.contains('.') {
-                    version = Some(name_or_version);
-                } else {
-                    contract_name = Some(name_or_version);
-                    version = parts.next();
-                }
-            }
-        } else {
-            contract_name = Some(path_or_name);
-            version = parts.next();
-        }
-
-        let version = if let Some(version) = version {
-            Some(Version::parse(version).map_err(|e| fmt_err!("failed parsing version: {e}"))?)
-        } else {
-            None
-        };
-
-        // Use available artifacts list if present
-        if let Some(artifacts) = &config.available_artifacts {
-            let filtered = artifacts
-                .iter()
-                .filter(|(id, _)| {
-                    // name might be in the form of "Counter.0.8.23"
-                    let id_name = id.name.split('.').next().unwrap();
-
-                    if let Some(path) = &file {
-                        if !id.source.ends_with(path) {
-                            return false;
-                        }
-                    }
-                    if let Some(name) = contract_name {
-                        if id_name != name {
-                            return false;
-                        }
-                    }
-                    if let Some(ref version) = version {
-                        if id.version.minor != version.minor ||
-                            id.version.major != version.major ||
-                            id.version.patch != version.patch
-                        {
-                            return false;
-                        }
-                    }
-                    true
-                })
-                .collect::<Vec<_>>();
-
-            let artifact = match &filtered[..] {
-                [] => Err(fmt_err!("no matching artifact found")),
-                [artifact] => Ok(artifact),
-                filtered => {
-                    // If we find more than one artifact, we need to filter by contract type
-                    // depending on whether we are using the zkvm or evm
-                    filtered
-                        .iter()
-                        .find(|(id, _)| {
-                            let contract_type =
-                                dual_compiled_contracts.get_contract_type_by_artifact(id);
-                            match contract_type {
-                                Some(ContractType::ZK) => using_zk_vm,
-                                Some(ContractType::EVM) => !using_zk_vm,
-                                None => false,
-                            }
-                        })
-                        .or_else(|| {
-                            // If we know the current script/test contract solc version, try to
-                            // filter by it
-                            config.running_version.as_ref().and_then(|version| {
-                                filtered.iter().find(|(id, _)| id.version == *version)
-                            })
-                        })
-                        .ok_or_else(|| fmt_err!("multiple matching artifacts found"))
-                }
-            }?;
-
-            let maybe_bytecode = if deployed {
-                artifact.1.deployed_bytecode().cloned()
-            } else {
-                artifact.1.bytecode().cloned()
-            };
-
-            return maybe_bytecode
-                .ok_or_else(|| fmt_err!("no bytecode for contract; is it abstract or unlinked?"));
-        } else {
-            let path_in_artifacts =
-                match (file.map(|f| f.to_string_lossy().to_string()), contract_name) {
-                    (Some(file), Some(contract_name)) => {
-                        PathBuf::from(format!("{file}/{contract_name}.json"))
-                    }
-                    (None, Some(contract_name)) => {
-                        PathBuf::from(format!("{contract_name}.sol/{contract_name}.json"))
-                    }
-                    (Some(file), None) => {
-                        let name = file.replace(".sol", "");
-                        PathBuf::from(format!("{file}/{name}.json"))
-                    }
-                    _ => bail!("invalid artifact path"),
-                };
-
-            config.paths.artifacts.join(path_in_artifacts)
-        }
-    };
-
-    let path = config.ensure_path_allowed(path, FsAccessKind::Read)?;
-    let data = fs::read_to_string(path)?;
-    let artifact = serde_json::from_str::<ContractObject>(&data)?;
-    let maybe_bytecode = if deployed { artifact.deployed_bytecode } else { artifact.bytecode };
-    maybe_bytecode.ok_or_else(|| fmt_err!("no bytecode for contract; is it abstract or unlinked?"))
 }
 
 fn get_context(
