+++ crates/strategy/zksync/src/executor/runner.rs
@@ -7,15 +7,15 @@ use foundry_evm::{
     backend::{Backend, BackendResult, CowBackend},
     executors::{
         strategy::{
-            EvmExecutorStrategyRunner, ExecutorStrategy, ExecutorStrategyContext,
-            ExecutorStrategyExt, ExecutorStrategyRunner,
+            EvmExecutorStrategyRunner, ExecutorStrategyContext, ExecutorStrategyExt,
+            ExecutorStrategyRunner,
         },
         Executor,
     },
     inspectors::InspectorStack,
 };
 use foundry_zksync_compilers::dual_compiled_contracts::DualCompiledContracts;
-use foundry_zksync_core::{vm::ZkEnv, ZkTransactionMetadata, ZKSYNC_TRANSACTION_OTHER_FIELDS_KEY};
+use foundry_zksync_core::vm::ZkEnv;
 use revm::{
     primitives::{Env, EnvWithHandlerCfg, ResultAndState},
     Database,
@@ -24,29 +24,8 @@ use revm::{
 use crate::{
     backend::{ZksyncBackendStrategyBuilder, ZksyncInspectContext},
     cheatcode::ZksyncCheatcodeInspectorStrategyBuilder,
+    executor::{try_get_zksync_transaction_metadata, ZksyncExecutorStrategyContext},
 };
-
-/// Defines the context for [ZksyncExecutorStrategyRunner].
-#[derive(Debug, Default, Clone)]
-pub struct ZksyncExecutorStrategyContext {
-    transaction_context: Option<ZkTransactionMetadata>,
-    dual_compiled_contracts: DualCompiledContracts,
-    zk_env: ZkEnv,
-}
-
-impl ExecutorStrategyContext for ZksyncExecutorStrategyContext {
-    fn new_cloned(&self) -> Box<dyn ExecutorStrategyContext> {
-        Box::new(self.clone())
-    }
-
-    fn as_any_ref(&self) -> &dyn std::any::Any {
-        self
-    }
-
-    fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
-        self
-    }
-}
 
 /// Defines the [ExecutorStrategyRunner] strategy for ZKsync.
 #[derive(Debug, Default, Clone)]
@@ -224,30 +203,5 @@ impl ExecutorStrategyExt for ZksyncExecutorStrategyRunner {
         let ctx = get_context(ctx);
         let transaction_context = try_get_zksync_transaction_metadata(&other_fields);
         ctx.transaction_context = transaction_context;
-    }
-}
-
-pub fn try_get_zksync_transaction_metadata(
-    other_fields: &OtherFields,
-) -> Option<ZkTransactionMetadata> {
-    other_fields
-        .get_deserialized::<ZkTransactionMetadata>(ZKSYNC_TRANSACTION_OTHER_FIELDS_KEY)
-        .transpose()
-        .ok()
-        .flatten()
-}
-
-/// Create ZKsync strategy for [ExecutorStrategy].
-pub trait ZksyncExecutorStrategyBuilder {
-    /// Create new zksync strategy.
-    fn new_zksync() -> Self;
-}
-
-impl ZksyncExecutorStrategyBuilder for ExecutorStrategy {
-    fn new_zksync() -> Self {
-        Self {
-            runner: &ZksyncExecutorStrategyRunner,
-            context: Box::new(ZksyncExecutorStrategyContext::default()),
-        }
     }
 }
